+++ app/src/main/java/nextstep/shoppingcart/model/Product.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart.data.model
+package nextstep.shoppingcart.model
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
