+++ app/src/main/java/nextstep/shoppingcart/ui/MainActivity.kt
@@ -1,4 +1,4 @@
-package nextstep.shoppingcart
+package nextstep.shoppingcart.ui
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
