+++ app/src/main/java/nextstep/shoppingcart/ui/cart/CartActivity.kt
@@ -19,7 +19,7 @@ class CartActivity : ComponentActivity() {
                     color = MaterialTheme.colorScheme.background
                 ) {
                     ShoppingCartScreen(
-                        onClickBackButton = { finish() }
+                        onClickBack = { finish() },
                     )
                 }
             }
