+++ app/src/main/java/nextstep/shoppingcart/ui/product/detail/ProductDetailContent.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
-import nextstep.shoppingcart.data.model.Product
+import nextstep.shoppingcart.model.Product
 import nextstep.shoppingcart.ui.component.BlueBottomButton
 import nextstep.shoppingcart.ui.component.PriceLabel
 import nextstep.shoppingcart.ui.theme.BlackContent
